Create README.md
Hi I am Manav a young web developer 
This is a clone website which I have created using html and css.Give me your feedback which will help me a lot in improving my self. I will update this website time to time with adding some new features Note : currently this is not responsive but I promise you that coming days I will definitely make this website beautifully responsive.But for now this is only made for mobile so for better experience view this on your mobile 
Have a nice day.